feat(#699): add AsmMethod class
